# Backend/src/infrastructure/external/Docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Role": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"role_name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"permissions_id": {
						"type": "number",
						"format": "double"
					},
					"permission": {
						"$ref": "#/components/schemas/Permission"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"role_name",
					"description",
					"permissions_id",
					"created_at"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Permission": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/Role"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"description"
				],
				"type": "object",
				"additionalProperties": true
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"emergency_contact": {
						"type": "string",
						"nullable": true
					},
					"two_factor_auth": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"role_id": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"last_name",
					"password",
					"phone",
					"role_id",
					"created_at"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Pagination": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total_pages": {
						"type": "number",
						"format": "double"
					},
					"hasNextPage": {
						"type": "boolean"
					},
					"hasPrevPage": {
						"type": "boolean"
					}
				},
				"required": [
					"total",
					"page",
					"limit",
					"total_pages",
					"hasNextPage",
					"hasPrevPage"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"create_permission_dto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"description"
				],
				"type": "object",
				"additionalProperties": true
			},
			"register_user_dto": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"emergency_contact": {
						"type": "string",
						"nullable": true
					},
					"two_factor_auth": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"role_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"email",
					"name",
					"last_name",
					"password",
					"phone",
					"role_id"
				],
				"type": "object",
				"additionalProperties": true
			},
			"update_user_dto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Healthtech App",
		"description": "Healthtech App's Documentation",
		"version": "1.0.0",
		"contact": {
			"name": "Moises Armando Heredia",
			"email": "moisarm03@gmail.com"
		}
	},
	"paths": {
...
		"/permission": {
			"get": {
				"operationId": "Get_permissions",
				"responses": {
					"200": {
						"description": "Permissions retrieve succesfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {
													"properties": {
														"pagination": {
															"$ref": "#/components/schemas/Pagination"
														},
														"data": {
															"items": {
																"properties": {
																	"description": {
																		"type": "string"
																	},
																	"id": {
																		"type": "number",
																		"format": "double"
																	},
																	"name": {
																		"type": "string"
																	}
																},
																"required": [
																	"description",
																	"id",
																	"name"
																],
																"type": "object"
															},
															"type": "array"
														}
													},
													"required": [
														"pagination",
														"data"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {},
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/permission/{id}": {
			"get": {
				"operationId": "Search_permission",
				"responses": {
					"200": {
						"description": "Permission retrieve succesfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {
													"properties": {
														"description": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														},
														"name": {
															"type": "string"
														}
													},
													"required": [
														"description",
														"id",
														"name"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {},
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update_permission",
				"responses": {
					"200": {
						"description": "Permission updated succesfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {
													"properties": {
														"description": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														},
														"name": {
															"type": "string"
														}
													},
													"required": [
														"description",
														"id",
														"name"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {},
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/create_permission_dto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete_permission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/permission/new": {
			"post": {
				"operationId": "Create_permission",
				"responses": {
					"201": {
						"description": "Permission created succesfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {
													"properties": {
														"description": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														},
														"name": {
															"type": "string"
														}
													},
													"required": [
														"description",
														"id",
														"name"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {},
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/create_permission_dto"
							}
						}
					}
				}
			}
		},
...
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Registered",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {
													"$ref": "#/components/schemas/User"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {},
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/register_user_dto"
							}
						}
					}
				}
			}
		},
		"/auth/{email}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"201": {
						"description": "Info Updated",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {
													"$ref": "#/components/schemas/User"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {},
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/update_user_dto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete_user",
				"responses": {
					"201": {
						"description": "User Deleted",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}